deliting unnecessary code
