+++ config/web.php
@@ -6,7 +6,7 @@ $config = [
     'id' => 'photo_gallery',
     'name' => 'Фотогаллерея на Yii2 для CMG',
     'basePath' => dirname(__DIR__),
-    'defaultRoute' => 'photo/index',
+    'defaultRoute' => 'gallery/index',
     'bootstrap' => ['log'],
     'language' => 'ru-RU',
     'components' => [
