+++ messages/ru/pg.main.php
@@ -13,4 +13,5 @@ return [
     'Home'      => 'Главная',
     'email'     => 'Email',
     'User name' => 'Имя пользователя',
+    'Error'     => 'Произошла какая-то ошибка'
 ];
