+++ views/gallery/comment.php
