+++ views/gallery/view.php
