--- views/photo/_form.php
@@ -1,25 +1,0 @@
-<?php
-
-use yii\helpers\Html;
-use yii\widgets\ActiveForm;
-
-/* @var $this yii\web\View */
-/* @var $model app\models\PhotoForm */
-/* @var $form yii\widgets\ActiveForm */
-?>
-
-<div class="photo-form">
-
-    <?php $form = ActiveForm::begin(['options' => ['enctype' => 'multipart/form-data']]); ?>
-
-    <?= $form->field($model, 'file')->fileInput() ?>
-
-    <?= $form->field($model, 'description')->textInput(['maxlength' => 255]) ?>
-
-    <div class="form-group">
-        <?= Html::submitButton(Yii::t('pg.photo', 'Create'), ['class' => 'btn btn-success']) ?>
-    </div>
-
-    <?php ActiveForm::end(); ?>
-
-</div>
